Move check_for_gl_error to fix doc test error
